feat(ignore_is_terminal): Allow the is_terminal check to be overridden by environment variable. (#13)
Fixes: https://github.com/zkat/supports-color/issues/12

Currently the tests fail in environments where the output is written to something other than a TTY or a known CI environment. This is because the code makes a hard-coded call to `is_terminal` and disables colour output if that returns false.

This change introduces an environment variable `IGNORE_IS_TERMINAL` which overrides the `is_terminal` check. The tests which depend on colour output now set this environment variable so that they pass in such environments.